backend/requirements.txt
fastapi
uvicorn
supabase
sentry-sdk
python-dotenv